# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/products/{productId}": {
      "put": {
        "summary": "更新产品",
        "deprecated": false,
        "description": "更新产品",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "FARMER :FARMER\nPURCHASER :PURCHASER\nADMIN :ADMIN",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FARMER",
                "PURCHASER",
                "ADMIN"
              ]
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmerProductRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "查询产品详情",
        "deprecated": false,
        "description": "查询产品详情",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultFarmerProductVO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "farmerId": 0,
                    "farmName": "",
                    "categoryId": 0,
                    "categoryName": "",
                    "name": "",
                    "spec": "",
                    "unit": "",
                    "price": 0,
                    "minPurchase": 0,
                    "stock": 0,
                    "productionDate": "",
                    "shelfLife": "",
                    "productionMethod": "",
                    "originAreaId": 0,
                    "originAreaName": "",
                    "description": "",
                    "status": "",
                    "images": [
                      ""
                    ],
                    "createTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/product-category/{id}": {
      "get": {
        "summary": "获取分类详情",
        "deprecated": false,
        "description": "获取分类详情",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProductCategory"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "name": "",
                    "parentId": 0,
                    "attribute": {
                      "": ""
                    },
                    "status": "",
                    "createTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除分类",
        "deprecated": false,
        "description": "删除分类",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/statistics/product/{productId}": {
      "get": {
        "summary": "获取产品销售统计",
        "deprecated": false,
        "description": "获取产品销售统计",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProductSalesStats"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "productName": "",
                    "salesCount": 0,
                    "salesAmount": 0,
                    "reviewCount": 0,
                    "averageRating": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/product-images/product/{productId}": {
      "get": {
        "summary": "按产品ID查询图片列表",
        "deprecated": false,
        "description": "按产品ID查询图片列表",
        "tags": [],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListProductImage"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "productId": 0,
                      "imageUrl": "",
                      "imageType": "",
                      "sort": 0,
                      "createTime": ""
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/product-images/{id}": {
      "put": {
        "summary": "修改产品图片信息",
        "deprecated": false,
        "description": "修改产品图片信息",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductImage",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除产品图片",
        "deprecated": false,
        "description": "删除产品图片",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/product-images/{id}/sort": {
      "put": {
        "summary": "调整图片顺序",
        "deprecated": false,
        "description": "调整图片顺序",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/product-images/page": {
      "get": {
        "summary": "分页查询产品图片",
        "deprecated": false,
        "description": "分页查询产品图片",
        "tags": [],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "productId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "imageType",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageProductImage"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "productId": 0,
                        "imageUrl": "",
                        "imageType": "",
                        "sort": 0,
                        "createTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/stock-reservation/order/{orderId}": {
      "get": {
        "summary": "获取订单的库存预留记录",
        "deprecated": false,
        "description": "获取订单的库存预留记录",
        "tags": [],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultStockReservation"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "productId": 0,
                    "orderId": 0,
                    "reservedQuantity": 0,
                    "status": "",
                    "releaseReason": "",
                    "expiredTime": "",
                    "createTime": "",
                    "updateTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/stock-reservation/{reservationId}/confirm": {
      "put": {
        "summary": "确认库存预留（支付成功后调用）",
        "deprecated": false,
        "description": "确认库存预留（支付成功后调用）",
        "tags": [],
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "description": "",
            "type": "null"
          }
        }
      },
...
      "MapString": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string",
            "description": "需要进行排序的字段"
          },
          "asc": {
            "type": "boolean",
            "description": "是否正序排列，默认 true",
            "default": true
          }
        }
      },
...
      "FarmerProductRequest": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "spec": {
            "type": "string",
            "description": ""
          },
          "unit": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": "",
            "minimum": 0.01
          },
          "minPurchase": {
            "type": "integer",
            "description": "",
            "minimum": 1
          },
          "stock": {
            "type": "integer",
            "description": "",
            "minimum": 1
          },
          "productionDate": {
            "type": "string",
            "description": ""
          },
          "shelfLife": {
            "type": "string",
            "description": ""
          },
          "productionMethod": {
            "type": "string",
            "description": ""
          },
          "originAreaId": {
            "type": "integer",
            "description": ""
          },
          "description": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "categoryId",
          "name",
          "spec",
          "unit",
          "price",
          "minPurchase",
          "stock",
          "originAreaId"
        ]
      },
...
      "FarmerProductVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "farmerId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "farmName": {
            "type": "string",
            "description": "关联农场名称"
          },
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "categoryName": {
            "type": "string",
            "description": "关联分类名称"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "spec": {
            "type": "string",
            "description": ""
          },
          "unit": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "minPurchase": {
            "type": "integer",
            "description": ""
          },
          "stock": {
            "type": "integer",
            "description": ""
          },
          "productionDate": {
            "type": "string",
            "description": ""
          },
          "shelfLife": {
            "type": "string",
            "description": ""
          },
          "productionMethod": {
            "type": "string",
            "description": ""
          },
          "originAreaId": {
            "type": "integer",
            "description": ""
          },
          "originAreaName": {
            "type": "string",
            "description": "关联产地名称"
          },
          "description": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "",
            "enum": [
              "ON_SALE",
              "OFF_SALE"
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "产品图片列表"
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ResultFarmerProductVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/FarmerProductVO",
            "description": ""
          }
        }
      },
...
      "ProductSalesStats": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string",
            "description": "产品名称"
          },
          "salesCount": {
            "type": "integer",
            "description": "销售数量"
          },
          "salesAmount": {
            "type": "number",
            "description": "销售总额"
          },
          "reviewCount": {
            "type": "integer",
            "description": "评价数"
          },
          "averageRating": {
            "type": "number",
            "description": "平均评分"
          }
        }
      },
...
      "ResultProductSalesStats": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/ProductSalesStats",
            "description": ""
          }
        }
      },
...
      "ProductCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "parentId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "attribute": {
            "$ref": "#/components/schemas/MapString",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ProductImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "imageUrl": {
            "type": "string",
            "description": ""
          },
          "imageType": {
            "type": "string",
            "description": "",
            "enum": [
              "COVER",
              "PRODUCTION",
              "DETAIL"
            ]
          },
          "sort": {
            "type": "integer",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ResultProductCategory": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/ProductCategory",
            "description": ""
          }
        }
      },
      "ResultListProductImage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage",
              "description": "产品图片实体类"
            },
            "description": ""
          }
        }
      },
...
      "PageProductImage": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage",
              "description": "产品图片实体类"
            },
            "description": "查询数据列表",
            "default": "Collections.emptyList()"
          },
          "total": {
            "type": "integer",
            "description": "总数",
            "format": "int64",
            "default": 0
          },
          "size": {
            "type": "integer",
            "description": "每页显示条数，默认 10",
            "format": "int64",
            "default": 10
          },
          "current": {
            "type": "integer",
            "description": "当前页",
            "format": "int64",
            "default": 1
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": "排序字段信息",
            "default": "new ArrayList<>()"
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": "自动优化 COUNT SQL",
            "default": true
          },
          "searchCount": {
            "type": "boolean",
            "description": "是否进行 count 查询",
            "default": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": "{@link #optimizeJoinOfCountSql()}",
            "default": true
          },
          "maxLimit": {
            "type": "integer",
            "description": "单页分页条数限制",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": "countId"
          }
        }
      },
...
      "ResultPageProductImage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PageProductImage",
            "description": ""
          }
        }
      },
...
      "StockReservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "description": "产品ID",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "reservedQuantity": {
            "type": "integer",
            "description": "预留数量"
          },
          "status": {
            "type": "string",
            "description": "预留状态：reserved(已预留) released(已释放) expired(已过期)"
          },
          "releaseReason": {
            "type": "string",
            "description": "释放原因"
          },
          "expiredTime": {
            "type": "string",
            "description": "预留过期时间"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间"
          }
        }
      },
      "ResultStockReservation": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/StockReservation",
            "description": ""
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}